requirements.txt
Flask==2.2.2
aiohttp==3.8.1
selenium==4.1.0
beautifulsoup4==4.10.0
apscheduler==3.9.1
pytz==2021.3
python-dotenv==0.19.2
logging==0.5.1.2